blog + login routes, access token creation

+++ backend/db/repository/blog.py
@@ -20,10 +20,12 @@ def list_blogs(db: Session):
     blogs = db.query(Blog).filter(Blog.is_active==True).all()
     return blogs
 
-def update_blog_by_id(id: int, blog: UpdateBlog, db: Session, author_id: int = 1):
+def update_blog_by_id(id: int, blog: UpdateBlog, db: Session, author_id: int = 1): # not the best because we return two different data types
     blog_in_db = db.query(Blog).filter(Blog.id==id).first()
     if not blog_in_db:
-        return
+        return {"error":f"Blog with id {id} does not exist."}
+    if not blog_in_db.author_id == author_id:
+        return {"error":f"Only the author can modify the blog."}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content 
     db.add(blog_in_db)
@@ -33,7 +35,9 @@ def update_blog_by_id(id: int, blog: UpdateBlog, db: Session, author_id: int = 1
 def delete_blog_by_id(id: int, db: Session, author_id: int):
     blog_in_db = db.query(Blog).filter(Blog.id==id)
     if not blog_in_db.first():
-        return {"error":f"could not find blog with the id {id}"}
+        return {"error":f"could not find blog with the id {id}."}
+    if not blog_in_db.first().author_id==author_id:
+        return {"error":f"Only the author can delete the blog."}
     blog_in_db.delete()
     db.commit()
     return {"msg":f"Deleted blog with id {id}"}
